+++ packages/auth-plugins/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.1](https://github.com/deephaven/web-client-ui/compare/v1.26.0...v1.26.1) (2026-07-21)
+
+**Note:** Version bump only for package @deephaven/auth-plugins
+
 ## [1.23.0](https://github.com/deephaven/web-client-ui/compare/v1.22.1...v1.23.0) (2026-06-25)
 
 **Note:** Version bump only for package @deephaven/auth-plugins